adds background image to mint and description screen
