+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-2.0.6/","dependencies":[]},{"name":"sqflite","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/sqflite-2.0.0+4/","dependencies":[]}],"android":[{"name":"path_provider","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-2.0.6/","dependencies":[]},{"name":"sqflite","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/sqflite-2.0.0+4/","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_macos-2.0.2/","dependencies":[]},{"name":"sqflite","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/sqflite-2.0.0+4/","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_linux-2.1.0/","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_windows-2.0.3/","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"path_provider","dependencies":["path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"sqflite","dependencies":[]}],"date_created":"2021-10-30 02:31:55.714166","version":"2.5.0"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-2.0.6/","dependencies":[]},{"name":"sqflite","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/sqflite-2.0.0+4/","dependencies":[]}],"android":[{"name":"path_provider","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-2.0.6/","dependencies":[]},{"name":"sqflite","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/sqflite-2.0.0+4/","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_macos-2.0.2/","dependencies":[]},{"name":"sqflite","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/sqflite-2.0.0+4/","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_linux-2.1.0/","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/mac/development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_windows-2.0.3/","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"path_provider","dependencies":["path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"sqflite","dependencies":[]}],"date_created":"2021-11-02 08:02:02.683956","version":"2.5.0"}
