+++ lib/widgets/pylons_round_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 class PylonsRoundButton extends StatelessWidget {
   final VoidCallback onPressed;
-  PylonsRoundButton({Key? key, required this.onPressed}) : super(key: key);
+  const PylonsRoundButton({Key? key, required this.onPressed}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +13,7 @@ class PylonsRoundButton extends StatelessWidget {
       child: const Icon(Icons.arrow_forward_ios, color: Colors.white),
       style: ElevatedButton.styleFrom(
         shape: const CircleBorder(),
-        padding: const EdgeInsets.all(20),
+        padding: const EdgeInsets.all(15),
         primary: kBlue,
         shadowColor: kBlue,
         elevation: 10
