+++ pubspec.yaml
@@ -39,6 +39,8 @@ dependencies:
 
   cached_network_image: ^3.1.0
 
+  flutter_svg: ^0.23.0+1
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -62,8 +64,10 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-assets:
-  - assets/icons/
+  assets:
+    - assets/icons/
+    - assets/svg/
+    - assets/images/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
